Added typehinting to GameMap, plus minor bugfixes and refactoring
Previously, none of the methods in the GameMap class had typehinting. Given
that this is the module that most other developers will be using, it's
probably best that typehinting is available for readability. This has now
been added.

A minor fix has been made to the assigning of the eligible tile indexes for
the second city, along with some other minor changes.